refactor(appkit-ui): rename variants dir to ui-variants
Match the appkit plugin dir name for cross-package consistency. Files and the
public API are unchanged; only the internal directory moves.
